grader/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-functions==1.21.3
certifi==2024.12.14
charset-normalizer==3.4.1
deprecation==2.1.0
distro==1.9.0
fastjsonschema==2.21.1
frozenlist==1.5.0
gotrue==2.11.3
groq==0.16.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
multidict==6.1.0
nbformat==5.10.4
packaging==24.2
platformdirs==4.3.6
postgrest==0.19.3
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.3.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
